requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
asgiref==3.7.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2024.2.2
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.5
daphne==4.1.0
Django==5.0.2
django-cache-lock==0.2.5
django-redis==5.4.0
djangorestframework==3.14.0
dnspython==2.6.1
frozenlist==1.4.1
hiredis==2.3.2
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
msgpack==1.0.7
multidict==6.0.5
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pymongo==4.6.2
pyOpenSSL==24.0.0
pytz==2024.1
redis==5.0.1
requests==2.31.0
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==23.10.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
yarl==1.9.4
zope.interface==6.2